requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
blis==0.7.5
bpemb==0.3.3
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
conllu==4.4.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
entrypoints==0.3
filelock==3.4.0
flair==0.10
flatbuffers==2.0
fonttools==4.28.1
frozenlist==1.2.0
fsspec==2021.11.0
ftfy==6.0.3
future==0.18.2
gast==0.4.0
gdown==3.12.2
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
hdbscan==0.8.27
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
Janome==0.4.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
libclang==12.0.0
llvmlite==0.37.0
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.8.0
mpld3==0.3
multidict==5.2.0
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
pyDeprecate==0.3.1
Pygments==2.10.0
Pympler==0.9
pynndescent==0.5.5
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
segtok==1.5.10
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.95
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlitedict==1.7.0
srsly==2.4.2
streamlit==1.2.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.5
typer==0.4.0
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
umap-learn==0.5.2
urllib3==1.26.7
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
Wikipedia-API==0.5.4
wordcloud==1.8.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
